nits, renaming and moving files here and there
